+++ Moya/Moya.swift
@@ -76,7 +76,7 @@ public protocol MoyaTarget {
     var baseURL: NSURL { get }
     var path: String { get }
     var method: Moya.Method { get }
-    var parameters: [String: AnyObject] { get }
+    var parameters: [String: AnyObject]? { get }
     var sampleData: NSData { get }
 }